Fix the overload which prints ExtMachInst in X86.
--HG--
extra : convert_revision : 2ef8ee71999f36b09270ba9526c2846beda65051

+++ src/arch/x86/types.hh
@@ -184,8 +184,8 @@ namespace X86ISA
                            "prefixA = %#x,\n\t\tprefixB = %#x\n\t},\n\t"
                      "modRM = %#x,\n\tsib = %#x,\n\t"
                      "immediate = %#x,\n\tdisplacement = %#x\n}\n",
-                     emi.legacy, (uint8_t)emi.rex,
-                     emi.opcode.num, emi.opcode.op,
+                     (uint8_t)emi.legacy, (uint8_t)emi.rex,
+                     emi.opcode.num, (uint8_t)emi.opcode.op,
                      emi.opcode.prefixA, emi.opcode.prefixB,
                      (uint8_t)emi.modRM, (uint8_t)emi.sib,
                      emi.immediate, emi.displacement);
